Cleanup imports and move color into utils.
